Refactor whistle power as Int
Having it as Int will make it easier
to arithmetically modify zombie life
with whistle power.

+++ miniMancera/Model/OptionWhistlesVsZombies/Whistle/MOptionWhistlesVsZombiesWhistleBlue.swift
@@ -1,9 +1,9 @@
-import UIKit
+import Foundation
 
 class MOptionWhistlesVsZombiesWhistleBlue:MOptionWhistlesVsZombiesWhistleProtocol
 {
     private let kSpeed:TimeInterval = 20
-    private let kPower:CGFloat = 4
+    private let kPower:Int = 4
     
     var speed:TimeInterval
     {
@@ -13,7 +13,7 @@ class MOptionWhistlesVsZombiesWhistleBlue:MOptionWhistlesVsZombiesWhistleProtoco
         }
     }
     
-    var power:CGFloat
+    var power:Int
     {
         get
         {
